Move impeller::Comparable to //impeller/base.
Just some minor cleanup. Fixes https://github.com/flutter/flutter/issues/98686.
