+++ base/BUILD.gn
@@ -10,6 +10,8 @@ impeller_component("base") {
     "allocation.h",
     "backend_cast.h",
     "base.h",
+    "comparable.cc",
+    "comparable.h",
     "config.h",
     "promise.cc",
     "promise.h",
