+++ base/comparable.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "impeller/renderer/comparable.h"
+#include "impeller/base/comparable.h"
 
 namespace impeller {
