+++ renderer/backend/metal/shader_library_mtl.h
@@ -12,7 +12,7 @@
 #include <unordered_map>
 
 #include "flutter/fml/macros.h"
-#include "impeller/renderer/comparable.h"
+#include "impeller/base/comparable.h"
 #include "impeller/renderer/shader_library.h"
 
 namespace impeller {
