+++ renderer/shader_function.h
@@ -6,7 +6,7 @@
 
 #include "flutter/fml/hash_combine.h"
 #include "flutter/fml/macros.h"
-#include "impeller/renderer/comparable.h"
+#include "impeller/base/comparable.h"
 #include "impeller/renderer/shader_types.h"
 
 namespace impeller {
