+++ typographer/font.h
@@ -8,7 +8,7 @@
 #include <optional>
 
 #include "flutter/fml/macros.h"
-#include "impeller/renderer/comparable.h"
+#include "impeller/base/comparable.h"
 #include "impeller/typographer/glyph.h"
 #include "impeller/typographer/typeface.h"
